Split the logic into a scheduler and a processor for MeasurementProtocol
The scheduler `BQToMeasurementProtocol` is responsible for adding to the
queue jobs for the processor `BQToMeasurementProtocolProcessor`.

This enables parallelism for faster execution (1 million hits in about 6
minutes, 10 million hits in about an hour).

We hardcode the bigquery BQ_BATCH_SIZE to 50,000 which should perform in
less than 5 minutes in normal conditions. In case of slowniness we will
implement a robust retry system later on with memcache.