Flatten BQ rows with repeated fields & structures
In order to send Enhanced Ecommerce MP hits with several impression
lists and/or with a purchased products list, we need to create BQ tables
with rows containing nested arrays of structures. When we send these
rows to Google Ananlytics with MP hits, we need to flatten repeated and
nested fileds to Enhanced Ecommerce parameters.