Add a minimalistic worker for Cloud Function usage
Using Cloud Function basically gives the possibility to CRMint to run
arbitrary code without handling the technical complexity and security
concerns.

This worker defines a simple contract with the Webhook/CloudFunction
with 2 parameters `input_uri` and `output_uri`. It expects the function
to transform the csv file and store the output in another file.

The Webhook is responsible to fail the HTTP request if it cannot
proceed.

By default Cloud Functions have a timeout of 1 minute (can be extended
to 9 minutes).