Run tests for jobs and controller in two separate environments
When running jobs tests we were using both controller and jobs requirements.txt files to install dependencies. This led to clashes in dependencies and would mask bugs due to different versions of librairies used in the two apps.

To mitigate this we will run tests in two separate environments with only their dependencies (to mimick what happens on the App Engine side, each app is separate with their own dependencies).

NB: The common tests should be run in the two environments.
PiperOrigin-RevId: 462135781