Implement a new command to output the CRMint UI url
This command is necessary to separate the concerns between our `crmint cloud setup`, `crmint cloud migrate` and other commands.

We also implemented a waiting loop that will ping the UI url to notify the user when it is ready to be opened in the browser.

The Cloud Shell tutorial has been updated accordingly.

PiperOrigin-RevId: 495408716